Add crate support

+++ Models/IEntry.cs
@@ -8,7 +8,7 @@ namespace DieselBundleViewer.Models
     {
         string Name { get; }
         string Type { get; }
-        uint Size { get; }
+        ulong Size { get; }
         string EntryPath { get; }
         FolderEntry Parent { get; set; }
